Switched to ExternalInternalGradientCalculator
Passing all three RooGradMinimizer tests now! #11

This commit also contains a whole lot of debugging prints and some remaining code from many tries of getting the AnalyticalGradientCalculator to work (i.e. get every floating point bit exactly right). In the end, it turned out that getting that to work would involve rewriting almost all of Minuit2's classes into long double precision versions. The ExternalInternalGradientCalculator approach just assumes that the external gradient calculator does things in Minuit2-internal parameter space, which means many int2ext and ext2int steps are no longer necessary. Since these were the main cause of the precision loss, this problem is now gone.

+++ math/minuit2/inc/Minuit2/SinParameterTransformation.h
@@ -31,13 +31,13 @@ namespace ROOT {
 
       ~SinParameterTransformation() {}
 
-      double Int2ext(double Value, double Upper, double Lower) const;
-      double Ext2int(double Value, double Upper, double Lower,
+      long double Int2ext(long double Value, long double Upper, long double Lower) const;
+      long double Ext2int(long double Value, long double Upper, long double Lower,
                      const MnMachinePrecision&) const;
-      double DInt2Ext(double Value, double Upper, double Lower) const;
+      long double DInt2Ext(long double Value, long double Upper, long double Lower) const;
 
-      double D2Int2Ext(double Value, double Upper, double Lower) const;
-      double GStepInt2Ext(double Value, double Upper, double Lower) const;
+      long double D2Int2Ext(long double Value, long double Upper, long double Lower) const;
+      long double GStepInt2Ext(long double Value, long double Upper, long double Lower) const;
 
     private:
