+++ math/minuit2/src/SinParameterTransformation.cxx
@@ -19,21 +19,21 @@ namespace ROOT {
 
 
 
-    double SinParameterTransformation::Int2ext(double Value, double Upper, double Lower) const {
+    long double SinParameterTransformation::Int2ext(long double Value, long double Upper, long double Lower) const {
       // transformation from  to internal (unlimited) to external values (limited by Lower/Upper )
       return Lower + 0.5*(Upper - Lower)*(sin(Value) + 1.);
     }
 
-    double SinParameterTransformation::Ext2int(double Value, double Upper, double Lower, const MnMachinePrecision& prec) const {
+    long double SinParameterTransformation::Ext2int(long double Value, long double Upper, long double Lower, const MnMachinePrecision& prec) const {
       // transformation from external (limited by Lower/Upper )  to internal (unlimited) values given the lower/upper limits
 
-      double piby2 = 2.*atan(1.);
-      double distnn = 8.*sqrt(prec.Eps2());
-      double vlimhi = piby2 - distnn;
-      double vlimlo = -piby2 + distnn;
+      long double piby2 = 2.*atan(1.);
+      long double distnn = 8.*sqrt(prec.Eps2());
+      long double vlimhi = piby2 - distnn;
+      long double vlimlo = -piby2 + distnn;
 
-      double yy = 2.*(Value - Lower)/(Upper - Lower) - 1.;
-      double yy2 = yy*yy;
+      long double yy = 2.*(Value - Lower)/(Upper - Lower) - 1.;
+      long double yy2 = yy*yy;
       if(yy2 > (1. - prec.Eps2())) {
         if(yy < 0.) {
           // Lower limit
@@ -50,17 +50,17 @@ namespace ROOT {
       }
     }
 
-    double SinParameterTransformation::DInt2Ext(double Value, double Upper, double Lower) const {
+    long double SinParameterTransformation::DInt2Ext(long double Value, long double Upper, long double Lower) const {
       // return the derivative of the transformation d Ext/ d Int
       return 0.5*((Upper - Lower)*cos(Value));
     }
 
-    double SinParameterTransformation::D2Int2Ext(double Value, double Upper, double Lower) const {
+    long double SinParameterTransformation::D2Int2Ext(long double Value, long double Upper, long double Lower) const {
       // return the second derivative of the transformation d^2 Ext/ {d Int}^2
       return 0.5 * ( (Lower - Upper) * sin(Value) );
     }
 
-    double SinParameterTransformation::GStepInt2Ext(double /*Value*/, double /*Upper*/, double /*Lower*/) const {
+    long double SinParameterTransformation::GStepInt2Ext(long double /*Value*/, long double /*Upper*/, long double /*Lower*/) const {
       return 1.;
     }
